refactor(mcp-server): split core.mjs into 7 responsibility-based modules
Split the monolithic ~3400-line core.mjs into focused sub-modules under
core/ while preserving backward compatibility via facade re-exports.

Modules created:
- core/constants.mjs — shared constants (CANONICAL_PREFIX, etc.)
- core/response.mjs — response helpers (buildJsonToolResponse, etc.)
- core/prefix.mjs — prefix/tag normalization and Levenshtein
- core/plugins.mjs — plugin runtime (normalize, validate, data sources)
- core/tokens.mjs — design token operations and payloads
- core/cem.mjs — CEM indexing, icons, a11y, related components
- core/register.mjs — all tool/resource/prompt registration (registerAll)

core.mjs remains as a thin facade with named re-exports + createMcpServer
orchestration. All 207 tests pass unchanged.

Closes #251
